day19 秒杀

+++ pinyougou_root/pinyougou_parent/pinyougou_dao/src/main/java/com/pinyougou/mapper/TbSeckillGoodsMapper.java
@@ -7,7 +7,7 @@ import org.apache.ibatis.annotations.Param;
 import java.util.List;
 
 public interface TbSeckillGoodsMapper {
-    long countByExample(TbSeckillGoodsExample example);
+    int countByExample(TbSeckillGoodsExample example);
 
     int deleteByExample(TbSeckillGoodsExample example);
